requirements.txt
ultralytics==8.2.76
python-dotenv==1.0.1
